Create Dockerfile, update gradle settings, and add Docker Compose
This commit introduces a Dockerfile for the bp_back service, with instructions to build an application jar and run it in an OpenJDK container. Included in this commit are changes to gradle settings to include the bp_back module, a new docker-compose file to orchestrate services including bp_back, and an update to the project group and version in build.gradle.kts. Seemingly unnecessary plugin and dependencies declarations have been removed.