Integrate MongoDB and refactor mapper classes
This commit introduces MongoDB in the storage layer and includes the necessary MongoDB Kotlin Coroutine driver in the dependencies. It creates new MongoDB specific classes including a connection handler, repository, and item model. The current storage, ItemStorage, has been updated to interact with the MongoDB via the repository. Item mappers to and from the Mongo specific item model were also created. In the process, the ItemGqlMapper was moved and renamed to reflect its role in the GraphQL model. Changes in API endpoint testing scripts and gradle dependencies updates are also included in this commit.