Refactor Items component and add delete item functionality
Wrapped the `subscribeToMore` method inside a `useEffect` hook in the Items component to better handle subscription and updates. Also added mutation for deleting an item in the frontend and handling it in the backend along with adjusting WebSocket settings for more robust handling of real-time updates. The order of item display was also reversed.